reorganize the operators to separate modules
