+++ src/benchmark/ssb_workload.cc
@@ -18,15 +18,15 @@
 #include "ssb_workload.h"
 
 #include "execution/execution_plan.h"
-#include "operators/aggregate.h"
-#include "operators/aggregate_options.h"
+#include "operators/aggregate/aggregate.h"
+#include "operators/aggregate/aggregate_options.h"
 #include "operators/fused/select_build_hash.h"
-#include "operators/hash_aggregate.h"
-#include "operators/join.h"
-#include "operators/join_graph.h"
-#include "operators/lip.h"
-#include "operators/predicate.h"
-#include "operators/select.h"
+#include "operators/aggregate/hash_aggregate.h"
+#include "operators/join/join.h"
+#include "operators/join/join_graph.h"
+#include "operators/join/lip.h"
+#include "operators/select/predicate.h"
+#include "operators/select/select.h"
 #include "scheduler/scheduler.h"
 #include "scheduler/scheduler_flags.h"
 #include "storage/util.h"
