+++ src/benchmark/ssb_workload_lip.cc
@@ -16,14 +16,14 @@
 // under the License.
 
 #include "execution/execution_plan.h"
-#include "operators/aggregate.h"
+#include "operators/aggregate/aggregate.h"
 #include "operators/fused/filter_join.h"
 #include "operators/fused/select_build_hash.h"
-#include "operators/join.h"
-#include "operators/join_graph.h"
-#include "operators/lip.h"
-#include "operators/predicate.h"
-#include "operators/select.h"
+#include "operators/join/join.h"
+#include "operators/join/join_graph.h"
+#include "operators/join/lip.h"
+#include "operators/select/predicate.h"
+#include "operators/select/select.h"
 #include "scheduler/scheduler.h"
 #include "ssb_workload.h"
 #include "storage/util.h"
