+++ src/operators/aggregate/aggregate_const.h
@@ -19,7 +19,7 @@
 #define HUSTLE_AGGREGATE_CONST_H
 
 #include "operators/operator.h"
-#include "base_aggregate.h"
+#include "operators/aggregate/base_aggregate.h"
 
 namespace hustle::operators {
