+++ src/operators/fused/select_build_hash.h
@@ -23,10 +23,10 @@
 
 #include <string>
 
-#include "operators/join.h"
+#include "operators/join/join.h"
 #include "operators/operator.h"
-#include "operators/predicate.h"
-#include "operators/select.h"
+#include "operators/select/predicate.h"
+#include "operators/select/select.h"
 #include "operators/utils/operator_result.h"
 #include "storage/block.h"
 #include "storage/table.h"
