+++ src/operators/tests/filter_join_test.cc
@@ -26,8 +26,8 @@
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
 #include "operators/fused/select_build_hash.h"
-#include "operators/join.h"
-#include "operators/select.h"
+#include "operators/join/join.h"
+#include "operators/select/select.h"
 #include "scheduler/scheduler.h"
 #include "storage/block.h"
 #include "storage/util.h"
