+++ src/resolver/cresolver.cc
@@ -24,8 +24,8 @@
 #include "execution/execution_plan.h"
 #include "operators/fused/filter_join.h"
 #include "operators/fused/select_build_hash.h"
-#include "operators/hash_aggregate.h"
-#include "operators/select.h"
+#include "operators/aggregate/hash_aggregate.h"
+#include "operators/select/select.h"
 #include "operators/utils/operator_result.h"
 #include "resolver/select_resolver.h"
 #include "scheduler/threading/synchronization_lock.h"
